+++ src/components/TodoModal.jsx
@@ -5,7 +5,7 @@ import { useDispatch } from "react-redux";
 import { AnimatePresence, motion } from "framer-motion";
 import toast from "react-hot-toast";
 import { format } from "date-fns";
-import { addTodo, updateTodo } from "slices/todoSlice";
+import { addTodo, updateTodo } from "redux/todoSlice.js";
 import styles from "styles/modules/modal.module.scss";
 import { Button } from "components/Button.jsx";
