+++ src/main.jsx
@@ -6,7 +6,7 @@ import "@fontsource/poppins/500.css";
 import "@fontsource/poppins/600.css";
 import "@fontsource/poppins/700.css";
 import "styles/GlobalStyles.css";
-import { store } from "app/store";
+import { store } from "redux/store.js";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
 	<Provider store={store}>
